Fix!: Rename package, class, artifact
Rename package, class, artifact in order to reflect project's relation
to Hibernate.

BREAKING CHANGE:
- Rename class from 'BaseJpaEntity' to 'BaseHibernateEntity'
- Move class to package 'io.github.akuniutka.data'
- Rename artifact to 'hibernate-base-entity'
