+++ pom.xml
@@ -12,7 +12,7 @@
     </parent>
 
     <groupId>io.github.akuniutka</groupId>
-    <artifactId>base-jpa-entity</artifactId>
+    <artifactId>hibernate-base-entity</artifactId>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
