+++ src/test/java/io/github/akuniutka/model/PhoneWithProperEqualsAndHashCode.java
@@ -1,6 +1,6 @@
 package io.github.akuniutka.model;
 
-import io.github.akuniutka.BaseJpaEntity;
+import io.github.akuniutka.data.BaseHibernateEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
 import lombok.Getter;
@@ -10,7 +10,7 @@ import lombok.Setter;
 @Table(name = "phones_proper")
 @Getter
 @Setter
-public class PhoneWithProperEqualsAndHashCode extends BaseJpaEntity {
+public class PhoneWithProperEqualsAndHashCode extends BaseHibernateEntity {
 
     private String number;
     private String type;
